refactor: 계층형 아키텍처를 헥사고날 아키텍처로 변경
