+++ src/main/java/org/example/remedy/application/like/LikeService.java
@@ -1,12 +1,12 @@
-package org.example.remedy.domain.like.service;
+package org.example.remedy.application.like;
 
 import lombok.RequiredArgsConstructor;
 
 import org.example.remedy.application.dropping.exception.DroppingNotFoundException;
 import org.example.remedy.domain.dropping.Dropping;
 import org.example.remedy.domain.dropping.DroppingRepository;
-import org.example.remedy.domain.like.domain.Like;
-import org.example.remedy.domain.like.repository.LikeRepository;
+import org.example.remedy.domain.like.Like;
+import org.example.remedy.infrastructure.persistence.like.LikeRepository;
 import org.example.remedy.application.user.exception.UserNotFoundException;
 import org.example.remedy.domain.user.User;
 import org.example.remedy.infrastructure.persistence.user.UserRepository;
